requirement.txt
scapy>=2.5.0
cryptography>=41.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
colorama>=0.4.6
